ncod/master/frontend/requirements.txt
# Web框架
pywebio>=1.8.0,<2.0.0
fastapi>=0.100.0,<1.0.0
uvicorn>=0.23.0,<1.0.0

# 数据库
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.11.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
redis>=5.0.0,<6.0.0
aioredis>=2.0.0,<3.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0

# 异步支持
aiohttp>=3.8.0,<4.0.0
asyncio>=3.4.3,<4.0.0

# 数据压缩
pako>=2.1.0,<3.0.0
zlib>=1.2.13,<2.0.0

# 安全
passlib>=1.7.4,<2.0.0
python-jose>=3.3.0,<4.0.0
bcrypt>=4.0.0,<5.0.0
pyotp>=2.8.0,<3.0.0
python-multipart>=0.0.6,<1.0.0

# 监控和指标
prometheus-client>=0.17.0,<1.0.0
psutil>=5.9.0,<6.0.0
paramiko>=3.3.0,<4.0.0

# 设备管理
easysnmp>=0.2.5,<1.0.0
docker>=6.1.0,<7.0.0
kubernetes>=26.1.0,<27.0.0
libvirt-python>=9.0.0,<10.0.0
netmiko>=4.1.2,<5.0.0
napalm>=4.0.0,<5.0.0
ncclient>=0.6.13,<1.0.0

# 云平台支持
boto3>=1.28.0,<2.0.0
python-novaclient>=18.3.0,<19.0.0
python-keystoneclient>=5.1.0,<6.0.0
python-neutronclient>=9.0.0,<10.0.0
openstacksdk>=1.0.0,<2.0.0

# 网络设备支持
f5-sdk>=3.0.21,<4.0.0
fortiosapi>=1.0.5,<2.0.0
pan-python>=0.16.0,<1.0.0

# 数据处理
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
plotly>=5.15.0,<6.0.0
networkx>=3.0.0,<4.0.0

# 工具
python-dateutil>=2.8.2,<3.0.0
pydantic>=2.0.0,<3.0.0
email-validator>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
PyYAML>=6.0.0,<7.0.0
requests>=2.31.0,<3.0.0
cryptography>=41.0.0,<42.0.0
pyOpenSSL>=23.2.0,<24.0.0

# 日志
loguru>=0.7.0,<1.0.0

# 测试
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0 